changes way that dropdown options are loaded
this is a bit of a temporary fix, but we load all of the dropdown
options on app load. This prevents a bug where if a dataset is
changed, but the scenario name doesn't change the model run
times don't get updated, which can lead to no rows being loaded
and the app crashing.

The dropdown_options reactive isn't currently updated, so when
new model results become available the lists wont change. This
will need to be changed at somepoint, perhaps by storing these
options in a cosmosdb account and polling those results regularly?
Or by subscribing to some event that notifies when batch has
completed a job