handles general los strategy in model
general LOS should only apply to rows which otherwise do not have a strategy applied to them.

this isn't possible to calculate before running the model, because it depends upon what mitigators the user provides values for.

we now figure out after assigning mitigators which rows to apply general los reduction to