ensures activity that is loaded is sorted
with the interpolated GAMs there is an implicit assumption that the data is sorted by age when we predict values.

however, this isn't always the case, the national data is randomly sorted, and the provider data is sorted but that might just be a happy little accident.

this ensures that the data is sorted so the predicitions are valid.

fixes #279