improve progress reporting
there is a tqdm callback which can be used to report progress by updating the metadata of the params file in queue.

but, currently this callback doesn't always report that all of the runs are completed.

this ensures that when we have finished running we update the metadata with the amount of model runs, signalling that the stage has been complete