fix issue with config tests
the tests were not properly mocking the dotenv values, which caused the test to behave unexpectedly if users had a .env file with actual values in.

refactored the code to use dotenv.dotenv_values instead: this gives us a dictionary of the values in the .env file, rather than appending to the environment variables.