refactors run
ensure we run the baseline first and outside of the multiprocessing pool. then, run all of the model iterations inside of the pool, finally returning the concatenation of the baseline and model results.

this ensures that the baseline will definitely, always, be the first (0th) item in the results list. if we ever change the multiprocessing loop that could become important if results are returned out of order.

it also adds a neatness to the multiprocessing loop, which will now have a nicely divisible number of iterations to perform per worker.