adds delivery episode in spell aggregation
had to refactor how we produce all aggregations, as this new one needed to be a filter, rather than a simple aggregation of columns.

the aggregate method is now inside of Model, not InpatientsModel etc. the model type specific aggregations are now in specific_aggregations method.

This has allowed us to use more complex logic before calling the get_agg method, such as filtering the processed results.